updated utility functions
